requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 25.1.1
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
blas=1.0=mkl
bottleneck=1.4.2=py310ha9d4c09_0
brotli-python=1.0.9=py310h6a678d5_9
bzip2=1.0.8=h5eee18b_6
ca-certificates=2025.8.3=hbd8a1cb_0
contourpy=1.3.1=py310hdb19cb5_0
cycler=0.11.0=pyhd3eb1b0_0
fonttools=4.55.3=py310h5eee18b_0
freetype=2.13.3=h4a9f257_0
intel-openmp=2023.1.0=hdb19cb5_46306
joblib=1.4.2=pypi_0
jpeg=9e=h5eee18b_3
kiwisolver=1.4.8=py310h6a678d5_0
lcms2=2.16=h92b89f2_1
ld_impl_linux-64=2.40=h12ee557_0
lerc=4.0.0=h6a678d5_0
libdeflate=1.22=h5eee18b_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.7.0=hde9077f_0
libuuid=1.41.5=h5eee18b_0
libwebp-base=1.3.2=h5eee18b_1
lz4-c=1.9.4=h6a678d5_1
matplotlib-base=3.10.0=py310hbfdbfaf_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py310h5eee18b_2
mkl_fft=1.3.11=py310h5eee18b_0
mkl_random=1.2.8=py310h1128e8f_0
ncurses=6.4=h6a678d5_0
numexpr=2.10.1=py310h3c60e43_0
numpy=2.0.1=py310h5f9d8c6_1
numpy-base=2.0.1=py310hb5e798b_1
openjpeg=2.5.2=h0d4d230_1
openssl=3.0.16=h5eee18b_0
packaging=24.2=py310h06a4308_0
pandas=2.2.3=py310h6a678d5_0
pillow=11.1.0=py310hac6e08b_1
pip=25.0=py310h06a4308_0
pwlf=2.5.1=pyhd8ed1ab_0
pydoe=0.3.8=py_1
pyparsing=3.2.0=py310h06a4308_0
python=3.10.16=he870216_1
python-dateutil=2.9.0post0=py310h06a4308_2
python-tzdata=2025.2=pyhd3eb1b0_0
pytz=2024.1=py310h06a4308_0
readline=8.2=h5eee18b_0
scikit-learn=1.6.1=pypi_0
scipy=1.15.2=pypi_0
seaborn=0.13.2=py310h06a4308_3
setuptools=75.8.0=py310h06a4308_0
six=1.17.0=py310h06a4308_0
sqlite=3.45.3=h5eee18b_0
tbb=2021.8.0=hdb19cb5_0
threadpoolctl=3.6.0=pypi_0
tk=8.6.14=h39e8969_0
tzdata=2025a=h04d1e81_0
unicodedata2=15.1.0=py310h5eee18b_1
wheel=0.45.1=py310h06a4308_0
xz=5.6.4=h5eee18b_1
zlib=1.2.13=h5eee18b_1
zstd=1.5.6=hc292b87_0
